+++ lib/search_provider.dart
@@ -2,8 +2,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'paginated_state.dart';
 
 // Searches items and provides them to PaginatedSearchController
-abstract class SearchProvider<T> {
+abstract class SearchProvider<T, Q> {
   /// perform a search, you can get the respective inputs of the search using [ref]
   /// e.g. filter, search text, etc... and customize the query based on the pagination [state]
-  Future<List<T>> performSearch(Ref ref, PaginatedState<T> state);
+  Future<List<T>> performSearch(Ref ref, PaginatedState<T, Q> state);
 }