+++ lib/widgets/paginated_sliver_grid_view.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'abstract_paginated_view.dart';
 
 // Helper class for showing items of a paginated search in a Sliver Grid View
-class PaginatedSliverGridView<T> extends AbstractPaginatedView<T> {
+class PaginatedSliverGridView<T, Q> extends AbstractPaginatedView<T, Q> {
   final SliverGridDelegate gridDelegate;
 
   const PaginatedSliverGridView({